Rename related js file
