+++ src/popup/tabs/history_tab.js
